Rung 3: fix AMP fp16/fp32 dtype crash in VQ dead-code reseed; robust Colab notebook
- tokenizer.py: force VQ codebook math to fp32 (`flat = z.reshape(...).float()`).
  Under GPU autocast `flat` was fp16, so the dead-code `self.codebook[dead] =
  flat[pick]` index_put dtype-mismatched the fp32 buffer and crashed Stage 1 at the
  first reseed. EMA-VQ in fp16 is numerically poor anyway. CPU smoke test never hit
  this (autocast off on CPU).
- stage1_pretrain.py: detach loss before float() in the log row (warning).
- rung3_colab.ipynb: public-repo clone (no token); install split with auto-restart so
  the numpy 2.x swap loads clean; every stage cell self-`cd`s so it survives the
  CPU->GPU runtime switch.
- requirements.txt: note torch/pillow/matplotlib for Rung 3.

+++ pong_counterfactual/cjepa_rung3/tokenizer.py
@@ -100,7 +100,11 @@ class VectorQuantizerEMA(nn.Module):
         """sem (B,M,dim) -> (quantized (B,M,dim), indices (B,M), commit loss,
         perplexity). EMA codebook update happens here when training."""
         z = self.proj_in(sem)                                  # (B,M,cd)
-        flat = z.reshape(-1, self.cfg.code_dim)
+        # force fp32 for ALL codebook ops: the buffers are fp32, and under AMP `z`
+        # is fp16 -> the EMA matmuls and especially the dead-code index_put
+        # (self.codebook[dead] = flat[pick]) would dtype-mismatch the fp32 buffers.
+        # EMA-VQ in fp16 is numerically poor regardless, so this is the right call.
+        flat = z.reshape(-1, self.cfg.code_dim).float()
         idx = self.quantize_indices(flat)                      # (B*M,)
         e = self.codebook[idx]
         if self.training:
